Reorganize page components
