+++ morgoth/base/static/js/components/pages/AddonDetailsPage.jsx
@@ -2,7 +2,7 @@ import React, { PropTypes as pt } from 'react';
 
 import Paper from 'material-ui/Paper';
 
-import AddonDetailsContainer from '../containers/AddonDetailsContainer';
+import AddonDetailsContainer from '../../containers/AddonDetailsContainer';
 
 
 function AddonDetailsPage(props) {
